@@ -1,22 +1,56 @@
 [project]
 name = "pymino"
-version = "1.3.4.7"
-description = "Un proyecto en Python para..."
-authors = [
-  { name = "DarkyAmagai", email = "darkyamagai@outlook.es" }
+description = "Easily create a bot for Amino Apps using a modern easy to use synchronous library."
+authors = [{ name = "DarkyAmagai", email = "darkyamagai@outlook.es" }]
+maintainers = [
+  { name = "ViktorSky", email = "viktorbotsprojects@gmail.com" },
 ]
 readme = "README.md"
-requires-python = ">=3.8"
-license = {text = "MIT"}
+requires-python = ">=3.9"
+license = "MIT"
+license-files = ["LICENSE"]
 classifiers = [
+  "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
-  "License :: OSI Approved :: MIT License",
-  "Operating System :: OS Independent"
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11",
+  "Programming Language :: Python :: 3.12",
+  "Programming Language :: Python :: 3.13",
+  "Topic :: Software Development :: Libraries :: Python Modules",
+  "Typing :: Typed",
 ]
+keywords = [
+  "amino",
+  "pymino",
+  "narvii",
+  "amino-api",
+  "narvii-bots",
+  "aminoapps",
+  "amino-bot",
+  "amino-bots",
+]
+dynamic = ["dependencies", "optional-dependencies", "version"]
 
 [project.urls]
 Homepage = "https://github.com/DarkyAmagai/pymino"
+Issues = "https://github.com/DarkyAmagai/pymino/issues"
 
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=77.0.3"]
 build-backend = "setuptools.build_meta"
+
+[tool.setuptools.dynamic]
+version = { attr = "pymino.__version__" }
+dependencies = { file = "requirements.txt" }
+optional-dependencies.dev = { file = "requirements-dev.txt" }
+
+[tool.setuptools.package-data]
+"*" = ["py.typed"]
+
+[tool.pyright]
+pythonVersion = "3.9"
+typeCheckingMode = "strict"
+strictListInference = true
+strictDictionaryInference = true
+strictSetInference = true
