@@ -1,560 +1,373 @@
-from typing import List, Union
+from collections.abc import Iterator
+from typing import Any, Optional, Union
+
+__all__ = ("Bubble", "BubbleList")
 
 
 class Bubble:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except AttributeError:
-            self.data = None
-    
-    def _check_bubble(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
     @property
-    @_check_bubble
     def ref_object_id(self) -> str:
         """Returns the ref object id of the bubble."""
-        return self.data.get("refObjectId")
+        return self.data.get("refObjectId", "")
 
     @property
-    @_check_bubble
     def created_time(self) -> str:
         """Returns the time the bubble was created."""
-        return self.data.get("createdTime")
+        return self.data.get("createdTime", "")
 
     @property
-    @_check_bubble
-    def item_basic_info(self) -> dict:
+    def item_basic_info(self) -> dict[str, Any]:
         """Returns the basic info of the bubble."""
-        return self.data.get("itemBasicInfo")
+        return self.data.get("itemBasicInfo") or {}
 
     @property
-    @_check_bubble
     def icon(self) -> str:
         """Returns the icon url of the bubble."""
-        if self.item_basic_info is None:
-            return None
-        else:
-            return self.item_basic_info.get("icon")
-        
+        return self.item_basic_info.get("icon", "")
+
     @property
-    @_check_bubble
     def name(self) -> str:
         """Returns the name of the bubble."""
-        if self.item_basic_info is None:
-            return None
-        else:
-            return self.item_basic_info.get("name")
-        
+        return self.item_basic_info.get("name", "")
+
     @property
-    @_check_bubble
-    def item_restriction_info(self) -> dict:
+    def item_restriction_info(self) -> dict[str, Any]:
         """Returns the restriction info of the bubble."""
-        return self.data.get("itemRestrictionInfo")
-
+        return self.data.get("itemRestrictionInfo") or {}
 
     @property
-    @_check_bubble
-    def owner_uid(self) -> Union[str, None]:
+    def owner_uid(self) -> Optional[str]:
         """Returns the owner uid of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("ownerUid")
-        
-    @property
-    @_check_bubble
-    def owner_type(self) -> Union[int, None]:
-        """Returns the owner type of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("ownerType")
-        
-    @property
-    @_check_bubble
-    def restrict_type(self) -> Union[int, None]:
-        """Returns the restict type of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("restrictType")
-        
-    @property
-    @_check_bubble
-    def restrict_value(self) -> Union[int, None]:
-        """Returns the restrict value of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("restrictValue")
-        
-    @property
-    @_check_bubble
-    def available_duration(self) -> Union[int, None]:
-        """Returns the available duration of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("availableDuration")
-        
-    @property
-    @_check_bubble
-    def discount_value(self) -> Union[int, None]:
-        """Returns the discount value of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("discountValue")
-        
-    @property
-    @_check_bubble
-    def discount_status(self) -> Union[int, None]:
-        """Returns the discount status of the bubble."""
-        if self.item_restriction_info is None:
-            return None
-        else:
-            return self.item_restriction_info.get("discountStatus")
-        
-    @property
-    @_check_bubble
-    def ref_object(self) -> dict:
-        """Returns the ref object of the bubble."""
-        return self.data.get("refObject")
+        return self.item_restriction_info.get("ownerUid")
 
     @property
-    @_check_bubble
+    def owner_type(self) -> int:
+        """Returns the owner type of the bubble."""
+        return self.item_restriction_info.get("ownerType", 0)
+
+    @property
+    def restrict_type(self) -> int:
+        """Returns the restict type of the bubble."""
+        return self.item_restriction_info.get("restrictType", 0)
+
+    @property
+    def restrict_value(self) -> Optional[int]:
+        """Returns the restrict value of the bubble."""
+        return self.item_restriction_info.get("restrictValue")
+
+    @property
+    def available_duration(self) -> Optional[str]:
+        """Returns the available duration of the bubble."""
+        return self.item_restriction_info.get("availableDuration")
+
+    @property
+    def discount_value(self) -> Optional[int]:
+        """Returns the discount value of the bubble."""
+        return self.item_restriction_info.get("discountValue")
+
+    @property
+    def discount_status(self) -> int:
+        """Returns the discount status of the bubble."""
+        return self.item_restriction_info.get("discountStatus", 0)
+
+    @property
+    def ref_object(self) -> dict[str, Any]:
+        """Returns the ref object of the bubble."""
+        return self.data.get("refObject") or {}
+
+    @property
     def is_global(self) -> bool:
         """Returns whether the bubble is globally available."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isGloballyAvailable")
-        
-    @property
-    @_check_bubble
-    def available_community_ids(self) -> list:
-        """Return the available community ids of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("availableNdcIds")
-        
-    @property
-    @_check_bubble
-    def bubble_type(self) -> int:
-        """Returns the bubble type of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("bubbleType")
-        
-    @property
-    @_check_bubble
-    def bubble_id(self) -> str:
-        """Returns the bubble id of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("bubbleId")
-        
-    @property
-    @_check_bubble
-    def background_image(self) -> str:
-        """Returns the background image of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("backgroundImage")
-        
-    @property
-    @_check_bubble
-    def status(self) -> int:
-        """Returns the status of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("status")
-        
-    @property
-    @_check_bubble
-    def is_new(self) -> bool:
-        """Returns whether the bubble is new."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isNew")
-        
-    @property
-    @_check_bubble
-    def bubble_name(self) -> str:
-        """Returns the name of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("name")
-        
-    @property
-    @_check_bubble
-    def banner_image(self) -> str:
-        """Returns the banner image of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("bannerImage")
-        
-    @property
-    @_check_bubble
-    def resource_url(self) -> str:
-        """Returns the resource url of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("resourceUrl")
-        
-    @property
-    @_check_bubble
-    def ownership_status(self) -> dict:
-        """Returns the ownership status of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("ownershipStatus")
-        
-    @property
-    @_check_bubble
-    def deletable(self) -> bool:
-        """Returns whether the bubble is deletable."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("deletable")
-        
-    @property
-    @_check_bubble
-    def config(self) -> dict:
-        """Returns the config of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("config")
-        
-    @property
-    @_check_bubble
-    def version(self) -> int:
-        """Returns the version of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("version")
-        
-    @property
-    @_check_bubble
-    def modified_time(self) -> str:
-        """Returns the modified time of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("modifiedTime")
-        
-    @property
-    @_check_bubble
-    def is_activated(self) -> bool:
-        """Returns whether the bubble is activated."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isActivated")
-        
-    @property
-    @_check_bubble
-    def cover_image(self) -> str:
-        """Returns the cover image of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("coverImage")
-        
-    @property
-    @_check_bubble
-    def extensions(self) -> dict:
-        """Returns the extensions of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("extensions")
-        
-    @property
-    @_check_bubble
-    def template_id(self) -> str:
-        """Returns the template id of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("templateId")
+        return self.ref_object.get("isGloballyAvailable", False)
 
     @property
-    @_check_bubble
+    def available_community_ids(self) -> list[int]:
+        """Return the available community ids of the bubble."""
+        return self.ref_object.get("availableNdcIds") or []
+
+    @property
+    def bubble_type(self) -> int:
+        """Returns the bubble type of the bubble."""
+        return self.ref_object.get("bubbleType", 0)
+
+    @property
+    def bubble_id(self) -> str:
+        """Returns the bubble id of the bubble."""
+        return self.ref_object.get("bubbleId", "")
+
+    @property
+    def background_image(self) -> Optional[str]:
+        """Returns the background image of the bubble."""
+        return self.ref_object.get("backgroundImage")
+
+    @property
+    def status(self) -> int:
+        """Returns the status of the bubble."""
+        return self.ref_object.get("status", 0)
+
+    @property
+    def is_new(self) -> bool:
+        """Returns whether the bubble is new."""
+        return self.ref_object.get("isNew", False)
+
+    @property
+    def bubble_name(self) -> str:
+        """Returns the name of the bubble."""
+        return self.ref_object.get("name", "")
+
+    @property
+    def banner_image(self) -> Optional[str]:
+        """Returns the banner image of the bubble."""
+        return self.ref_object.get("bannerImage")
+
+    @property
+    def resource_url(self) -> str:
+        """Returns the resource url of the bubble."""
+        return self.ref_object.get("resourceUrl", "")
+
+    @property
+    def ownership_status(self) -> int:
+        """Returns the ownership status of the bubble."""
+        return self.ref_object.get("ownershipStatus", 0)
+
+    @property
+    def deletable(self) -> bool:
+        """Returns whether the bubble is deletable."""
+        return self.ref_object.get("deletable", False)
+
+    @property
+    def config(self) -> dict[str, Any]:
+        """Returns the config of the bubble."""
+        return self.ref_object.get("config") or {}
+
+    @property
+    def version(self) -> int:
+        """Returns the version of the bubble."""
+        return self.ref_object.get("version", 0)
+
+    @property
+    def modified_time(self) -> Optional[str]:
+        """Returns the modified time of the bubble."""
+        return self.ref_object.get("modifiedTime")
+
+    @property
+    def is_activated(self) -> bool:
+        """Returns whether the bubble is activated."""
+        return self.ref_object.get("isActivated", False)
+
+    @property
+    def cover_image(self) -> Optional[str]:
+        """Returns the cover image of the bubble."""
+        return self.ref_object.get("coverImage")
+
+    @property
+    def extensions(self) -> dict[str, Any]:
+        """Returns the extensions of the bubble."""
+        return self.ref_object.get("extensions") or {}
+
+    @property
+    def template_id(self) -> int:
+        """Returns the template id of the bubble."""
+        return self.ref_object.get("templateId", 0)
+
+    @property
     def uid(self) -> str:
         """Returns the uid of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("uid")
-        
+        return self.ref_object.get("uid", "")
+
     @property
-    @_check_bubble
-    def md5(self) -> str:
+    def md5(self) -> Optional[str]:
         """Returns the md5 of the bubble."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("md5")
-        
+        return self.ref_object.get("md5")
+
     @property
-    @_check_bubble
     def ref_object_type(self) -> int:
         """Returns the ref object type of the bubble."""
-        if self.data is None:
-            return None
-        else:
-            return self.data.get("refObjectType")
-        
+        return self.data.get("refObjectType", 0)
 
 
 class BubbleList:
-    def __init__(self, data: dict):
-        try:
-            self.data = data.get("storeItemList")
-        except Exception:
-            self.data = None
+    def __init__(self, data: Union[list[dict[str, Any]], dict[str, Any]]) -> None:
+        if isinstance(data, dict):
+            data = list(data.get("storeItemList") or [])
+        self.data = data or []
 
-    def _check_bubbles(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
-    def __iterator__(self):
-        """
-        Iterator function to iterate over the bubbles in the bubble list.
-        """
-        for bubble in self.data:
-            yield Bubble(bubble)
+    def __iter__(self) -> Iterator[Bubble]:
+        return (Bubble(bubble) for bubble in self.data)
 
     @property
-    @_check_bubbles
-    def ref_object_id(self) -> List[str]:
+    def ref_object_id(self) -> list[str]:
         """Returns the ref object id of the bubble list."""
-        return [bubble.ref_object_id for bubble in self.__iterator__()]
-    
+        return [bubble.ref_object_id for bubble in self]
+
     @property
-    @_check_bubbles
-    def created_time(self) -> List[str]:
+    def created_time(self) -> list[str]:
         """Returns the created time of the bubble list."""
-        return [bubble.created_time for bubble in self.__iterator__()]
-    
+        return [bubble.created_time for bubble in self]
+
     @property
-    @_check_bubbles
-    def item_basic_info(self) -> List[dict]:
+    def item_basic_info(self) -> list[dict[str, Any]]:
         """Returns the item basic info of the bubble list."""
-        return [bubble.item_basic_info for bubble in self.__iterator__()]
-    
+        return [bubble.item_basic_info for bubble in self]
+
     @property
-    @_check_bubbles
-    def icon(self) -> List[str]:
+    def icon(self) -> list[str]:
         """Returns the icon of the bubble list."""
-        return [bubble.icon for bubble in self.__iterator__()]
-    
+        return [bubble.icon for bubble in self]
+
     @property
-    @_check_bubbles
-    def name(self) -> List[str]:
+    def name(self) -> list[str]:
         """Returns the name of the bubble list."""
-        return [bubble.name for bubble in self.__iterator__()]
-    
+        return [bubble.name for bubble in self]
+
     @property
-    @_check_bubbles
-    def item_restriction_info(self) -> List[dict]:
+    def item_restriction_info(self) -> list[dict[str, Any]]:
         """Returns the item restriction info of the bubble list."""
-        return [bubble.item_restriction_info for bubble in self.__iterator__()]
-    
+        return [bubble.item_restriction_info for bubble in self]
+
     @property
-    @_check_bubbles
-    def owner_uid(self) -> List[str]:
+    def owner_uid(self) -> list[Optional[str]]:
         """Returns the owner uid of the bubble list."""
-        return [bubble.owner_uid for bubble in self.__iterator__()]
-    
+        return [bubble.owner_uid for bubble in self]
+
     @property
-    @_check_bubbles
-    def owner_type(self) -> List[int]:
+    def owner_type(self) -> list[int]:
         """Returns the owner type of the bubble list."""
-        return [bubble.owner_type for bubble in self.__iterator__()]
-    
+        return [bubble.owner_type for bubble in self]
+
     @property
-    @_check_bubbles
-    def restrict_type(self) -> List[int]:
+    def restrict_type(self) -> list[int]:
         """Returns the restrict type of the bubble list."""
-        return [bubble.restrict_type for bubble in self.__iterator__()]
-    
+        return [bubble.restrict_type for bubble in self]
+
     @property
-    @_check_bubbles
-    def restrict_value(self) -> List[int]:
+    def restrict_value(self) -> list[Optional[int]]:
         """Returns the restrict value of the bubble list."""
-        return [bubble.restrict_value for bubble in self.__iterator__()]
-    
+        return [bubble.restrict_value for bubble in self]
+
     @property
-    @_check_bubbles
-    def available_duration(self) -> List[str]:
+    def available_duration(self) -> list[Optional[str]]:
         """Returns the available duration of the bubble list."""
-        return [bubble.available_duration for bubble in self.__iterator__()]
-    
+        return [bubble.available_duration for bubble in self]
+
     @property
-    @_check_bubbles
-    def discount_value(self) -> List[str]:
+    def discount_value(self) -> list[Optional[int]]:
         """Returns the discount value of the bubble list."""
-        return [bubble.discount_value for bubble in self.__iterator__()]
-    
+        return [bubble.discount_value for bubble in self]
+
     @property
-    @_check_bubbles
-    def discount_status(self) -> List[int]:
+    def discount_status(self) -> list[int]:
         """Returns the discount status of the bubble list."""
-        return [bubble.discount_status for bubble in self.__iterator__()]
-    
+        return [bubble.discount_status for bubble in self]
+
     @property
-    @_check_bubbles
-    def ref_object(self) -> List[dict]:
+    def ref_object(self) -> list[dict[str, Any]]:
         """Returns the ref object of the bubble list."""
-        return [bubble.ref_object for bubble in self.__iterator__()]
-    
+        return [bubble.ref_object for bubble in self]
+
     @property
-    @_check_bubbles
-    def is_global(self) -> List[str]:
+    def is_global(self) -> list[bool]:
         """Returns the is global of the bubble list."""
-        return [bubble.is_global for bubble in self.__iterator__()]
-    
+        return [bubble.is_global for bubble in self]
+
     @property
-    @_check_bubbles
-    def available_community_ids(self) -> List[int]:
+    def available_community_ids(self) -> list[list[int]]:
         """Returns the available community ids of the bubble list."""
-        return [bubble.available_community_ids for bubble in self.__iterator__()]
-    
+        return [bubble.available_community_ids for bubble in self]
+
     @property
-    @_check_bubbles
-    def bubble_type(self) -> List[int]:
+    def bubble_type(self) -> list[int]:
         """Returns the bubble type of the bubble list."""
-        return [bubble.bubble_type for bubble in self.__iterator__()]
-    
+        return [bubble.bubble_type for bubble in self]
+
     @property
-    @_check_bubbles
-    def bubble_id(self) -> List[str]:
+    def bubble_id(self) -> list[str]:
         """Returns the bubble id of the bubble list."""
-        return [bubble.bubble_id for bubble in self.__iterator__()]
-    
+        return [bubble.bubble_id for bubble in self]
+
     @property
-    @_check_bubbles
-    def background_image(self) -> List[str]:
+    def background_image(self) -> list[Optional[str]]:
         """Returns the background image of the bubble list."""
-        return [bubble.background_image for bubble in self.__iterator__()]
+        return [bubble.background_image for bubble in self]
 
     @property
-    @_check_bubbles
-    def status(self) -> List[int]:
+    def status(self) -> list[int]:
         """Returns the status of the bubble list."""
-        return [bubble.status for bubble in self.__iterator__()]
+        return [bubble.status for bubble in self]
 
     @property
-    @_check_bubbles
-    def is_new(self) -> List[str]:
+    def is_new(self) -> list[bool]:
         """Returns the is new of the bubble list."""
-        return [bubble.is_new for bubble in self.__iterator__()]
+        return [bubble.is_new for bubble in self]
 
     @property
-    @_check_bubbles
-    def name(self) -> List[str]:
-        """Returns the name of the bubble list."""
-        return [bubble.name for bubble in self.__iterator__()]
-
-    @property
-    @_check_bubbles
-    def banner_image(self) -> List[str]:
+    def banner_image(self) -> list[Optional[str]]:
         """Returns the banner image of the bubble list."""
-        return [bubble.banner_image for bubble in self.__iterator__()]
+        return [bubble.banner_image for bubble in self]
 
     @property
-    @_check_bubbles
-    def resource_url(self) -> List[str]:
+    def resource_url(self) -> list[str]:
         """Returns the resource url of the bubble list."""
-        return [bubble.resource_url for bubble in self.__iterator__()]
+        return [bubble.resource_url for bubble in self]
 
     @property
-    @_check_bubbles
-    def ownership_status(self) -> List[str]:
+    def ownership_status(self) -> list[int]:
         """Returns the ownership status of the bubble list."""
-        return [bubble.ownership_status for bubble in self.__iterator__()]
+        return [bubble.ownership_status for bubble in self]
 
     @property
-    @_check_bubbles
-    def deletable(self) -> List[str]:
+    def deletable(self) -> list[bool]:
         """Returns the deletable of the bubble list."""
-        return [bubble.deletable for bubble in self.__iterator__()]
+        return [bubble.deletable for bubble in self]
 
     @property
-    @_check_bubbles
-    def config(self) -> List[dict]:
+    def config(self) -> list[dict[str, Any]]:
         """Returns the config of the bubble list."""
-        return [bubble.config for bubble in self.__iterator__()]
+        return [bubble.config for bubble in self]
 
     @property
-    @_check_bubbles
-    def version(self) -> List[int]:
+    def version(self) -> list[int]:
         """Returns the version of the bubble list."""
-        return [bubble.version for bubble in self.__iterator__()]
+        return [bubble.version for bubble in self]
 
     @property
-    @_check_bubbles
-    def modified_time(self) -> List[str]:
+    def modified_time(self) -> list[Optional[str]]:
         """Returns the modified time of the bubble list."""
-        return [bubble.modified_time for bubble in self.__iterator__()]
+        return [bubble.modified_time for bubble in self]
 
     @property
-    @_check_bubbles
-    def is_activated(self) -> List[str]:
+    def is_activated(self) -> list[bool]:
         """Returns the is activated of the bubble list."""
-        return [bubble.is_activated for bubble in self.__iterator__()]
+        return [bubble.is_activated for bubble in self]
 
     @property
-    @_check_bubbles
-    def cover_image(self) -> List[str]:
+    def cover_image(self) -> list[Optional[str]]:
         """Returns the cover image of the bubble list."""
-        return [bubble.cover_image for bubble in self.__iterator__()]
+        return [bubble.cover_image for bubble in self]
 
     @property
-    @_check_bubbles
-    def extensions(self) -> List[str]:
+    def extensions(self) -> list[dict[str, Any]]:
         """Returns the extensions of the bubble list."""
-        return [bubble.extensions for bubble in self.__iterator__()]
+        return [bubble.extensions for bubble in self]
 
     @property
-    @_check_bubbles
-    def template_id(self) -> List[str]:
+    def template_id(self) -> list[int]:
         """Returns the template id of the bubble list."""
-        return [bubble.template_id for bubble in self.__iterator__()]
+        return [bubble.template_id for bubble in self]
 
     @property
-    @_check_bubbles
-    def uid(self) -> List[str]:
+    def uid(self) -> list[str]:
         """Returns the uid of the bubble list."""
-        return [bubble.uid for bubble in self.__iterator__()]
+        return [bubble.uid for bubble in self]
 
     @property
-    @_check_bubbles
-    def md5(self) -> List[str]:
+    def md5(self) -> list[Optional[str]]:
         """Returns the md5 of the bubble list."""
-        return [bubble.md5 for bubble in self.__iterator__()]
-    
+        return [bubble.md5 for bubble in self]
 
-    def json(self) -> dict:
+    def json(self) -> list[dict[str, Any]]:
         """Returns the json of the bubble list."""
         return self.data
