@@ -1,1020 +1,765 @@
-from typing import List, Union
+from collections.abc import Iterator
+from typing import Any, Optional, Union
+
+__all__ = (
+    "CommunitySticker",
+    "CommunityStickerList",
+    "OriginalCommunity",
+    "Sticker",
+    "StickerAuthor",
+    "StickerExtensions",
+    "StickerList",
+)
 
 
 class StickerAuthor:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
-    def _check_author(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
+    def __bool__(self) -> bool:
+        return bool(self.uid)
+
     @property
-    @_check_author
     def is_global(self) -> bool:
         """Returns True if the author is global, False if not."""
-        return self.data.get("isGlobal")
-    
+        return self.data.get("isGlobal", False)
+
     @property
-    @_check_author
     def comId(self) -> int:
         """Returns the author's community ID."""
-        return self.data.get("ndcId")
-    
+        return self.data.get("ndcId", 0)
+
     @property
-    @_check_author
     def role(self) -> int:
         """Returns the author's role."""
-        return self.data.get("role")
-    
+        return self.data.get("role", 0)
+
     @property
-    @_check_author
     def is_nickname_verified(self) -> bool:
         """Returns True if the author's nickname is verified, False if not."""
-        return self.data.get("isNicknameVerified")
-    
+        return self.data.get("isNicknameVerified", False)
+
     @property
-    @_check_author
     def following_status(self) -> int:
         """Returns the author's following status."""
-        return self.data.get("followingStatus")
-    
+        return self.data.get("followingStatus", 0)
+
     @property
-    @_check_author
     def followers(self) -> int:
         """Returns the author's follower count."""
-        return self.data.get("membersCount")
-    
+        return self.data.get("membersCount", 0)
+
     @property
-    @_check_author
     def membership_status(self) -> int:
         """Returns the author's membership status."""
-        return self.data.get("membershipStatus")
-    
+        return self.data.get("membershipStatus", 0)
+
     @property
-    @_check_author
     def status(self) -> int:
         """Returns the author's status."""
-        return self.data.get("status")
-    
+        return self.data.get("status", 0)
+
     @property
-    @_check_author
     def account_membership_status(self) -> int:
         """Returns the author's account membership status."""
-        return self.data.get("accountMembershipStatus")
-    
+        return self.data.get("accountMembershipStatus", 0)
+
     @property
-    @_check_author
     def reputation(self) -> int:
         """Returns the author's reputation."""
-        return self.data.get("reputation")
-    
+        return self.data.get("reputation", 0)
+
     @property
-    @_check_author
     def level(self) -> int:
         """Returns the author's level."""
-        return self.data.get("level")
-    
+        return self.data.get("level", 0)
+
     @property
-    @_check_author
-    def icon(self) -> str:
+    def icon(self) -> Optional[str]:
         """Returns the author's icon."""
         return self.data.get("icon")
-    
+
     @property
-    @_check_author
     def uid(self) -> str:
         """Returns the author's UID."""
-        return self.data.get("uid")
-    
+        return self.data.get("uid", "")
+
     @property
-    @_check_author
     def nickname(self) -> str:
         """Returns the author's nickname."""
-        return self.data.get("nickname")
-    
-    def json(self) -> dict:
+        return self.data.get("nickname", "")
+
+    def json(self) -> dict[str, Any]:
         """Returns the author's data as a dictionary."""
         return self.data
 
 
 class OriginalCommunity:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
-    def _check_community(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
+    def __bool__(self) -> bool:
+        return bool(self.comId)
+
     @property
-    @_check_community
     def status(self) -> int:
         """Returns the community's status."""
-        return self.data.get("status")
-    
+        return self.data.get("status", 0)
+
     @property
-    @_check_community
     def icon(self) -> str:
         """Returns the community's icon."""
-        return self.data.get("icon")
-    
+        return self.data.get("icon", "")
+
     @property
-    @_check_community
     def endpoint(self) -> str:
         """Returns the community's endpoint."""
-        return self.data.get("endpoint")
-    
+        return self.data.get("endpoint", "")
+
     @property
-    @_check_community
     def name(self) -> str:
         """Returns the community's name."""
-        return self.data.get("name")
-    
+        return self.data.get("name", "")
+
     @property
-    @_check_community
     def comId(self) -> int:
         """Returns the community's ndc id."""
-        return self.data.get("ndcId")
-    
-    def json(self) -> dict:
+        return self.data.get("ndcId", 0)
+
+    def json(self) -> dict[str, Any]:
         """Returns the community's json."""
         return self.data
 
 
 class StickerExtensions:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
-    def _check_extension(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
+    def __bool__(self) -> bool:
+        return bool(self.data)
+
     @property
-    @_check_extension
     def icon_source_sticker_id(self) -> str:
         """Returns the sticker's icon source sticker id."""
-        return self.data.get("iconSourceStickerId")
-    
+        return self.data.get("iconSourceStickerId", "")
+
     @property
-    @_check_extension
     def original_author(self) -> StickerAuthor:
         """Returns the sticker's original author."""
-        return StickerAuthor(self.data.get("originalAuthor"))
-    
+        return StickerAuthor(self.data.get("originalAuthor") or {})
+
     @property
-    @_check_extension
     def original_community(self) -> OriginalCommunity:
         """Returns the sticker's original community."""
-        return OriginalCommunity(self.data.get("originalCommunity"))
-    
-    def json(self) -> dict:
+        return OriginalCommunity(self.data.get("originalCommunity") or {})
+
+    def json(self) -> dict[str, Any]:
         """Returns the sticker's extensions as a dictionary."""
         return self.data
 
 
 class Sticker:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
-    def _check_sticker(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
+    def __bool__(self) -> bool:
+        return bool(self.created_time)
 
     @property
-    @_check_sticker
     def ref_object_type(self) -> int:
         """Returns the sticker's reference object type."""
-        return self.data.get("refObjectType")
+        return self.data.get("refObjectType", 0)
 
     @property
-    @_check_sticker
     def ref_object_id(self) -> str:
         """Returns the sticker's reference object ID."""
-        return self.data.get("refObjectId")
+        return self.data.get("refObjectId", "")
 
     @property
-    @_check_sticker
     def created_time(self) -> str:
         """Returns the sticker's creation time."""
-        return self.data.get("createdTime")
+        return self.data.get("createdTime", "")
 
     @property
-    @_check_sticker
-    def item_basic_info(self) -> dict:
+    def item_basic_info(self) -> dict[str, Any]:
         """Returns the sticker's basic info."""
-        return self.data.get("itemBasicInfo")
+        return self.data.get("itemBasicInfo") or {}
 
     @property
-    @_check_sticker
     def icon(self) -> str:
         """Returns the sticker's icon."""
-        return self.item_basic_info.get("icon")
+        return self.item_basic_info.get("icon", "")
 
     @property
-    @_check_sticker
     def name(self) -> str:
         """Returns the sticker's name."""
-        return self.item_basic_info.get("name")
+        return self.item_basic_info.get("name", "")
 
     @property
-    @_check_sticker
-    def item_restriction_info(self) -> dict:
+    def item_restriction_info(self) -> dict[str, Any]:
         """Returns the sticker's restriction info."""
-        return self.data.get("itemRestrictionInfo")
+        return self.data.get("itemRestrictionInfo") or {}
 
     @property
-    @_check_sticker
     def discount_value(self) -> Union[int, None]:
         """Returns the sticker's discount value."""
         return self.item_restriction_info.get("discountValue")
 
     @property
-    @_check_sticker
     def discount_status(self) -> int:
         """Returns the sticker's discount status."""
-        return self.item_restriction_info.get("discountStatus")
+        return self.item_restriction_info.get("discountStatus", 0)
 
     @property
-    @_check_sticker
-    def owner_uid(self) -> Union[str, None]:
+    def owner_uid(self) -> Optional[str]:
         """Returns the sticker's owner UID."""
         return self.item_restriction_info.get("ownerUid")
 
     @property
-    @_check_sticker
     def owner_type(self) -> int:
         """Returns the sticker's owner type."""
-        return self.item_restriction_info.get("ownerType")
+        return self.item_restriction_info.get("ownerType", 0)
 
     @property
-    @_check_sticker
     def restrict_type(self) -> int:
         """Returns the sticker's restriction type."""
-        return self.item_restriction_info.get("restrictType")
+        return self.item_restriction_info.get("restrictType", 0)
 
     @property
-    @_check_sticker
-    def restrict_value(self) -> int:
+    def restrict_value(self) -> Optional[int]:
         """Returns the sticker's restriction value."""
         return self.item_restriction_info.get("restrictValue")
 
     @property
-    @_check_sticker
-    def available_duration(self) -> Union[int, None]:
+    def available_duration(self) -> Optional[int]:
         """Returns the sticker's available duration."""
         return self.item_restriction_info.get("availableDuration")
 
     @property
-    @_check_sticker
-    def ref_object(self) -> dict:
+    def ref_object(self) -> dict[str, Any]:
         """Returns the sticker's reference object."""
-        return self.data.get("refObject")
+        return self.data.get("refObject") or {}
 
     @property
-    @_check_sticker
-    def modified_time(self) -> str:
+    def modified_time(self) -> Optional[str]:
         """Returns the sticker's modification time."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("modifiedTime")
+        return self.ref_object.get("modifiedTime")
 
     @property
-    @_check_sticker
-    def ownership_status(self) -> Union[int, None]:
+    def ownership_status(self) -> Optional[int]:
         """Returns the sticker's ownership status."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("ownershipStatus")
+        return self.ref_object.get("ownershipStatus")
 
     @property
-    @_check_sticker
+    def is_owned(self) -> bool:
+        """Returns the sticker list's ownership status."""
+        return self.ref_object.get("isOwned", False)
+
+    @property
     def used_count(self) -> int:
         """Returns the sticker's used count."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("usedCount")
+        return self.ref_object.get("usedCount", 0)
 
     @property
-    @_check_sticker
-    def extensions(self) -> Union[dict, None]:
+    def extensions(self) -> dict[str, Any]:
         """Returns the sticker's extensions."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("extensions")
+        return self.ref_object.get("extensions") or {}
 
     @property
-    @_check_sticker
-    def available_community_ids(self) -> list:
+    def available_community_ids(self) -> list[int]:
         """Returns the sticker's available community IDs."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("availableNdcIds")
+        return self.ref_object.get("availableNdcIds") or []
 
     @property
-    @_check_sticker
     def status(self) -> int:
         """Returns the sticker's status."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("status")
+        return self.ref_object.get("status", 0)
 
     @property
-    @_check_sticker
     def author(self) -> StickerAuthor:
         """Returns the sticker's author."""
-        return StickerAuthor(self.ref_object.get("author"))
+        return StickerAuthor(self.ref_object.get("author") or {})
 
     @property
-    @_check_sticker
     def is_new(self) -> bool:
         """Returns the sticker's new status."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isNew")
+        return self.ref_object.get("isNew", False)
 
     @property
-    @_check_sticker
     def stickers_count(self) -> int:
         """Returns the sticker's count."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("stickersCount")
+        return self.ref_object.get("stickersCount", 0)
 
     @property
-    @_check_sticker
-    def restriction_info(self) -> dict:
+    def restriction_info(self) -> dict[str, Any]:
         """Returns the sticker's restriction info."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("restrictionInfo")
+        return self.ref_object.get("restrictionInfo") or {}
 
     @property
-    @_check_sticker
-    def collection_id(self) -> str:
+    def collection_id(self) -> Optional[str]:
         """Returns the sticker's collection ID."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("collectionId")
+        return self.ref_object.get("collectionId")
 
     @property
-    @_check_sticker
     def is_activated(self) -> bool:
         """Returns the sticker's activation status."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isActivated")
+        return self.ref_object.get("isActivated", False)
 
     @property
-    @_check_sticker
     def collection_type(self) -> int:
         """Returns the sticker's collection type."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("collectionType")
+        return self.ref_object.get("collectionType", 0)
 
     @property
-    @_check_sticker
-    def uid(self) -> str:
+    def uid(self) -> Optional[str]:
         """Returns the sticker's UID."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("uid")
+        return self.ref_object.get("uid")
 
     @property
-    @_check_sticker
     def small_icon(self) -> str:
         """Returns the sticker's small icon."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("smallIcon")
+        return self.ref_object.get("smallIcon", "")
 
     @property
-    @_check_sticker
-    def description(self) -> str:
+    def description(self) -> Optional[str]:
         """Returns the sticker's description."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("description")
+        return self.ref_object.get("description")
 
     @property
-    @_check_sticker
     def is_globally_available(self) -> bool:
         """Returns the sticker's global availability status."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("isGloballyAvailable")
+        return self.ref_object.get("isGloballyAvailable", False)
 
     @property
-    @_check_sticker
     def banner_url(self) -> str:
         """Returns the sticker's banner URL."""
-        if self.ref_object is None:
-            return None
-        else:
-            return self.ref_object.get("bannerUrl")
-        
-    def json(self) -> dict:
+        return self.ref_object.get("bannerUrl", "")
+
+    def json(self) -> dict[str, Any]:
         """Returns the sticker's data in JSON format."""
         return self.data
-    
+
 
 class StickerList:
-    def __init__(self, data: dict):
-        try:
-            self.data = data.get("storeItemList")
-        except Exception:
-            self.data = None
+    def __init__(self, data: Union[list[dict[str, Any]], dict[str, Any]]):
+        if isinstance(data, dict):
+            data = list(data.get("storeItemList") or [])
+        self.data = data
 
-    def _check_stickers(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
-    def __iterator__(self):
-        """
-        Iterator function to iterate over the sticker list.
-        """
-        for sticker in self.data:
-            yield Sticker(sticker)
+    def __iter__(self) -> Iterator[Sticker]:
+        return (Sticker(sticker) for sticker in self.data)
 
     @property
-    @_check_stickers
-    def ref_object_type(self) -> List[int]:
+    def ref_object_type(self) -> list[int]:
         """Returns the sticker list's ref object type."""
-        return [sticker.ref_object_type for sticker in self.__iterator__()]
-    
+        return [sticker.ref_object_type for sticker in self]
+
     @property
-    @_check_stickers
-    def ref_object_id(self) -> List[str]:
+    def ref_object_id(self) -> list[str]:
         """Returns the sticker list's ref object ID."""
-        return [sticker.ref_object_id for sticker in self.__iterator__()]
-    
+        return [sticker.ref_object_id for sticker in self]
+
     @property
-    @_check_stickers
-    def created_time(self) -> List[str]:
+    def created_time(self) -> list[str]:
         """Returns the sticker list's creation time."""
-        return [sticker.created_time for sticker in self.__iterator__()]
-    
+        return [sticker.created_time for sticker in self]
+
     @property
-    @_check_stickers
-    def item_basic_info(self) -> List[dict]:
+    def item_basic_info(self) -> list[dict[str, Any]]:
         """Returns the sticker list's basic info."""
-        return [sticker.item_basic_info for sticker in self.__iterator__()]
-    
+        return [sticker.item_basic_info for sticker in self]
+
     @property
-    @_check_stickers
-    def icon(self) -> List[str]:
+    def icon(self) -> list[str]:
         """Returns the sticker list's icon."""
-        return [sticker.icon for sticker in self.__iterator__()]
-    
+        return [sticker.icon for sticker in self]
+
     @property
-    @_check_stickers
-    def name(self) -> List[str]:
+    def name(self) -> list[str]:
         """Returns the sticker list's name."""
-        return [sticker.name for sticker in self.__iterator__()]
-    
+        return [sticker.name for sticker in self]
+
     @property
-    @_check_stickers
-    def item_restriction_info(self) -> List[dict]:
+    def item_restriction_info(self) -> list[dict[str, Any]]:
         """Returns the sticker list's restriction info."""
-        return [sticker.item_restriction_info for sticker in self.__iterator__()]
-    
+        return [sticker.item_restriction_info for sticker in self]
+
     @property
-    @_check_stickers
-    def discount_value(self) -> List[str]:
+    def discount_value(self) -> list[Optional[int]]:
         """Returns the sticker list's discount value."""
-        return [sticker.discount_value for sticker in self.__iterator__()]
-    
+        return [sticker.discount_value for sticker in self]
+
     @property
-    @_check_stickers
-    def discount_status(self) -> List[int]:
+    def discount_status(self) -> list[int]:
         """Returns the sticker list's discount status."""
-        return [sticker.discount_status for sticker in self.__iterator__()]
-    
+        return [sticker.discount_status for sticker in self]
+
     @property
-    @_check_stickers
-    def owner_uid(self) -> List[str]:
+    def owner_uid(self) -> list[Optional[str]]:
         """Returns the sticker list's owner UID."""
-        return [sticker.owner_uid for sticker in self.__iterator__()]
-    
+        return [sticker.owner_uid for sticker in self]
+
     @property
-    @_check_stickers
-    def owner_type(self) -> List[int]:
+    def owner_type(self) -> list[int]:
         """Returns the sticker list's owner type."""
-        return [sticker.owner_type for sticker in self.__iterator__()]
-    
+        return [sticker.owner_type for sticker in self]
+
     @property
-    @_check_stickers
-    def restrict_type(self) -> List[int]:
+    def restrict_type(self) -> list[int]:
         """Returns the sticker list's restrict type."""
-        return [sticker.restrict_type for sticker in self.__iterator__()]
-    
+        return [sticker.restrict_type for sticker in self]
+
     @property
-    @_check_stickers
-    def restrict_value(self) -> List[int]:
+    def restrict_value(self) -> list[Optional[int]]:
         """Returns the sticker list's restrict value."""
-        return [sticker.restrict_value for sticker in self.__iterator__()]
-    
+        return [sticker.restrict_value for sticker in self]
+
     @property
-    @_check_stickers
-    def available_duration(self) -> List[str]:
+    def available_duration(self) -> list[Optional[int]]:
         """Returns the sticker list's available duration."""
-        return [sticker.available_duration for sticker in self.__iterator__()]
-    
+        return [sticker.available_duration for sticker in self]
+
     @property
-    @_check_stickers
-    def ownership_status(self) -> List[int]:
+    def ownership_status(self) -> list[Optional[int]]:
         """Returns the sticker list's ownership status."""
-        return [sticker.ownership_status for sticker in self.__iterator__()]
-    
+        return [sticker.ownership_status for sticker in self]
+
     @property
-    @_check_stickers
-    def is_owned(self) -> List[bool]:
+    def is_owned(self) -> list[bool]:
         """Returns the sticker list's ownership status."""
-        return [sticker.is_owned for sticker in self.__iterator__()]
-    
+        return [sticker.is_owned for sticker in self]
+
     @property
-    @_check_stickers
-    def is_new(self) -> List[bool]:
+    def is_new(self) -> list[bool]:
         """Returns the sticker list's new status."""
-        return [sticker.is_new for sticker in self.__iterator__()]
-    
+        return [sticker.is_new for sticker in self]
+
     @property
-    @_check_stickers
-    def is_activated(self) -> List[bool]:
+    def is_activated(self) -> list[bool]:
         """Returns the sticker list's activation status."""
-        return [sticker.is_activated for sticker in self.__iterator__()]
-    
+        return [sticker.is_activated for sticker in self]
+
     @property
-    @_check_stickers
-    def collection_id(self) -> List[str]:
+    def collection_id(self) -> list[Optional[str]]:
         """Returns the sticker list's collection ID."""
-        return [sticker.collection_id for sticker in self.__iterator__()]
-    
+        return [sticker.collection_id for sticker in self]
+
     @property
-    @_check_stickers
-    def collection_type(self) -> List[int]:
+    def collection_type(self) -> list[int]:
         """Returns the sticker list's collection type."""
-        return [sticker.collection_type for sticker in self.__iterator__()]
-    
+        return [sticker.collection_type for sticker in self]
+
     @property
-    @_check_stickers
-    def uid(self) -> List[str]:
+    def uid(self) -> list[Optional[str]]:
         """Returns the sticker list's UID."""
-        return [sticker.uid for sticker in self.__iterator__()]
-    
+        return [sticker.uid for sticker in self]
+
     @property
-    @_check_stickers
-    def small_icon(self) -> List[str]:
+    def small_icon(self) -> list[str]:
         """Returns the sticker list's small icon."""
-        return [sticker.small_icon for sticker in self.__iterator__()]
-    
+        return [sticker.small_icon for sticker in self]
+
     @property
-    @_check_stickers
-    def description(self) -> List[str]:
+    def description(self) -> list[Optional[str]]:
         """Returns the sticker list's description."""
-        return [sticker.description for sticker in self.__iterator__()]
-    
+        return [sticker.description for sticker in self]
+
     @property
-    @_check_stickers
-    def is_globally_available(self) -> List[bool]:
+    def is_globally_available(self) -> list[bool]:
         """Returns the sticker list's global availability."""
-        return [sticker.is_globally_available for sticker in self.__iterator__()]
-    
+        return [sticker.is_globally_available for sticker in self]
+
     @property
-    @_check_stickers
-    def banner_url(self) -> List[str]:
+    def banner_url(self) -> list[str]:
         """Returns the sticker list's banner URL."""
-        return [sticker.banner_url for sticker in self.__iterator__()]
-    
-    def json(self) -> dict:
+        return [sticker.banner_url for sticker in self]
+
+    def json(self) -> list[dict[str, Any]]:
         """Returns the sticker list's JSON data."""
         return self.data
 
 
 class CommunitySticker:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data = data
 
-    def _check_sticker(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-    
     @property
-    @_check_sticker
     def status(self) -> int:
         """Returns the sticker's status."""
-        return self.data.get("status")
-    
+        return self.data.get("status", 0)
+
     @property
-    @_check_sticker
     def is_activated(self) -> bool:
         """Returns True if the sticker is activated, False otherwise."""
-        return self.data.get("isActivated")
-    
+        return self.data.get("isActivated", False)
+
     @property
-    @_check_sticker
     def collection_type(self) -> int:
         """Returns the sticker's collection type."""
-        return self.data.get("collectionType")
-    
+        return self.data.get("collectionType", 0)
+
     @property
-    @_check_sticker
     def is_new(self) -> bool:
         """Returns True if the sticker is new, False otherwise."""
-        return self.data.get("isNew")
-    
+        return self.data.get("isNew", False)
+
     @property
-    @_check_sticker
     def banner_url(self) -> str:
         """Returns the sticker's banner url."""
-        return self.data.get("bannerUrl")
+        return self.data.get("bannerUrl", "")
 
     @property
-    @_check_sticker
     def is_owned(self) -> bool:
         """Returns True if the sticker is owned, False otherwise."""
-        return self.data.get("isOwned")
-    
+        return self.data.get("isOwned", False)
+
     @property
-    @_check_sticker
     def used_count(self) -> int:
         """Returns the sticker's used count."""
-        return self.data.get("usedCount")
-    
+        return self.data.get("usedCount", False)
+
     @property
-    @_check_sticker
-    def available_community_ids(self) -> list:
+    def available_community_ids(self) -> list[int]:
         """Returns a list of available community ids."""
-        return self.data.get("availableNdcIds")
-    
+        return self.data.get("availableNdcIds") or []
+
     @property
-    @_check_sticker
     def icon(self) -> str:
         """Returns the sticker's icon url."""
-        return self.data.get("icon")
-    
+        return self.data.get("icon", "")
+
     @property
-    @_check_sticker
     def name(self) -> str:
         """Returns the sticker's name."""
-        return self.data.get("name")
+        return self.data.get("name", "")
 
     @property
-    @_check_sticker
     def collection_id(self) -> str:
         """Returns the sticker's collection id."""
-        return self.data.get("collectionId")
+        return self.data.get("collectionId", "")
 
     @property
-    @_check_sticker
-    def description(self) -> str:
+    def description(self) -> Optional[str]:
         """Returns the sticker's description."""
         return self.data.get("description")
 
     @property
-    @_check_sticker
     def author(self) -> StickerAuthor:
         """Returns the sticker's author."""
-        return StickerAuthor(self.data.get("author"))
+        return StickerAuthor(self.data.get("author") or {})
 
     @property
-    @_check_sticker
-    def ownership_info(self) -> dict:
+    def ownership_info(self) -> dict[str, Any]:
         """Returns the sticker's ownership info."""
-        return self.data.get("ownershipInfo")
+        return self.data.get("ownershipInfo") or {}
 
     @property
-    @_check_sticker
     def is_auto_renew(self) -> bool:
         """Returns True if the sticker is auto renew, False otherwise."""
-        if self.owner_info is None:
-            return None
-        return self.owner_info.get("isAutoRenew")
+        return self.ownership_info.get("isAutoRenew", False)
 
     @property
-    @_check_sticker
-    def expired_time(self) -> str:
+    def expired_time(self) -> Optional[str]:
         """Returns the sticker's expired time."""
-        if self.owner_info is None:
-            return None
-        return self.owner_info.get("expiredTime")
+        return self.ownership_info.get("expiredTime")
 
     @property
-    @_check_sticker
     def extensions(self) -> StickerExtensions:
         """Returns the sticker's extensions."""
-        return StickerExtensions(self.data.get("extensions"))
+        return StickerExtensions(self.data.get("extensions") or {})
 
     @property
-    @_check_sticker
     def created_time(self) -> str:
         """Returns the sticker's created time."""
-        return self.data.get("createdTime")
+        return self.data.get("createdTime", "")
 
     @property
-    @_check_sticker
     def is_globally_available(self) -> bool:
         """Returns True if the sticker is globally available, False otherwise."""
-        return self.data.get("isGloballyAvailable")
-
-    
-    @property   
-    @_check_sticker
-    def restriction_info(self) -> dict:
-        """Returns the sticker's restriction info."""
-        return self.data.get("restrictionInfo")
+        return self.data.get("isGloballyAvailable", False)
 
     @property
-    @_check_sticker
+    def restriction_info(self) -> dict[str, Any]:
+        """Returns the sticker's restriction info."""
+        return self.data.get("restrictionInfo") or {}
+
+    @property
     def discount_status(self) -> int:
         """Returns the sticker's discount status."""
-        if self.restriction_info is None:
-            return None
-        return self.restriction_info.get("discountStatus")
+        return self.restriction_info.get("discountStatus", 0)
 
     @property
-    @_check_sticker
-    def owner_uid(self) -> str:
+    def owner_uid(self) -> Optional[str]:
         """Returns the sticker's owner uid."""
-        if self.restriction_info is None:
-            return None
         return self.restriction_info.get("ownerUid")
 
     @property
-    @_check_sticker
     def owner_type(self) -> int:
         """Returns the sticker's owner type."""
-        if self.restriction_info is None:
-            return None
-        return self.restriction_info.get("ownerType")
+        return self.restriction_info.get("ownerType", 0)
 
     @property
-    @_check_sticker
     def restrict_type(self) -> int:
         """Returns the sticker's restrict type."""
-        if self.restriction_info is None:
-            return None
-        return self.restriction_info.get("restrictType")
+        return self.restriction_info.get("restrictType", 0)
 
     @property
-    @_check_sticker
     def restrict_value(self) -> int:
         """Returns the sticker's restrict value."""
-        if self.restriction_info is None:
-            return None
-        return self.restriction_info.get("restrictValue")
+        return self.restriction_info.get("restrictValue", 0)
 
     @property
-    @_check_sticker
-    def available_duration(self) -> str:
+    def available_duration(self) -> Optional[str]:
         """Returns the sticker's available duration."""
-        if self.restriction_info is None:
-            return None
         return self.restriction_info.get("availableDuration")
 
     @property
-    @_check_sticker
-    def discount_value(self) -> str:
+    def discount_value(self) -> Optional[int]:
         """Returns the sticker's discount value."""
-        if self.restriction_info is None:
-            return None
         return self.restriction_info.get("discountValue")
 
-    def json(self) -> dict:
+    def json(self) -> dict[str, Any]:
         """Returns the sticker's data as a dict."""
         return self.data
-    
+
 
 class CommunityStickerList:
-    def __init__(self, data: dict) -> None:
-        try:
-            self.data = data.get("stickerCollectionList")
-        except Exception:
-            self.data = None
+    def __init__(self, data: dict[str, Any]) -> None:
+        self.data: list[dict[str, Any]] = data.get("stickerCollectionList") or []
 
-    def _check_sticker(F):
-        def wrapper(*args, **kwargs):
-            return None if args[0].data is None else F(*args, **kwargs)
-        return wrapper
-
-    def __iterator__(self):
-        """
-        Iterator function to iterate over the stickers in the sticker list.
-        """
-        for sticker in self.data:
-            yield CommunitySticker(sticker)
+    def __iter__(self) -> Iterator[CommunitySticker]:
+        return (CommunitySticker(sticker) for sticker in self.data)
 
     @property
-    @_check_sticker
-    def status(self) -> List[int]:
+    def status(self) -> list[int]:
         """Returns the sticker's status."""
-        return [sticker.status for sticker in self.__iterator__()]
-    
+        return [sticker.status for sticker in self]
+
     @property
-    @_check_sticker
-    def is_activated(self) -> bool:
+    def is_activated(self) -> list[bool]:
         """Returns True if the sticker is activated, False otherwise."""
-        return [sticker.is_activated for sticker in self.__iterator__()]
-    
+        return [sticker.is_activated for sticker in self]
+
     @property
-    @_check_sticker
-    def collection_type(self) -> int:
+    def collection_type(self) -> list[int]:
         """Returns the sticker's collection type."""
-        return [sticker.collection_type for sticker in self.__iterator__()]
-    
+        return [sticker.collection_type for sticker in self]
+
     @property
-    @_check_sticker
-    def is_new(self) -> bool:
+    def is_new(self) -> list[bool]:
         """Returns True if the sticker is new, False otherwise."""
-        return [sticker.is_new for sticker in self.__iterator__()]
-    
+        return [sticker.is_new for sticker in self]
+
     @property
-    @_check_sticker
-    def banner_url(self) -> str:
+    def banner_url(self) -> list[str]:
         """Returns the sticker's banner url."""
-        return [sticker.banner_url for sticker in self.__iterator__()]
+        return [sticker.banner_url for sticker in self]
 
     @property
-    @_check_sticker
-    def is_owned(self) -> bool:
+    def is_owned(self) -> list[bool]:
         """Returns True if the sticker is owned, False otherwise."""
-        return [sticker.is_owned for sticker in self.__iterator__()]
-    
+        return [sticker.is_owned for sticker in self]
+
     @property
-    @_check_sticker
-    def used_count(self) -> int:
+    def used_count(self) -> list[int]:
         """Returns the sticker's used count."""
-        return [sticker.used_count for sticker in self.__iterator__()]
-    
+        return [sticker.used_count for sticker in self]
+
     @property
-    @_check_sticker
-    def available_community_ids(self) -> list:
+    def available_community_ids(self) -> list[list[int]]:
         """Returns a list of available community ids."""
-        return [sticker.available_community_ids for sticker in self.__iterator__()]
-    
+        return [sticker.available_community_ids for sticker in self]
+
     @property
-    @_check_sticker
-    def icon(self) -> str:
+    def icon(self) -> list[str]:
         """Returns the sticker's icon url."""
-        return [sticker.icon for sticker in self.__iterator__()]
-    
+        return [sticker.icon for sticker in self]
+
     @property
-    @_check_sticker
-    def name(self) -> str:
+    def name(self) -> list[str]:
         """Returns the sticker's name."""
-        return [sticker.name for sticker in self.__iterator__()]
+        return [sticker.name for sticker in self]
 
     @property
-    @_check_sticker
-    def collection_id(self) -> str:
+    def collection_id(self) -> list[str]:
         """Returns the sticker's collection id."""
-        return [sticker.collection_id for sticker in self.__iterator__()]
+        return [sticker.collection_id for sticker in self]
 
     @property
-    @_check_sticker
-    def description(self) -> str:
+    def description(self) -> list[Optional[str]]:
         """Returns the sticker's description."""
-        return [sticker.description for sticker in self.__iterator__()]
+        return [sticker.description for sticker in self]
 
     @property
-    @_check_sticker
-    def ownership_info(self) -> dict:
+    def ownership_info(self) -> list[dict[str, Any]]:
         """Returns the sticker's ownership info."""
-        return [sticker.ownership_info for sticker in self.__iterator__()]
+        return [sticker.ownership_info for sticker in self]
 
     @property
-    @_check_sticker
-    def is_auto_renew(self) -> bool:
+    def is_auto_renew(self) -> list[bool]:
         """Returns True if the sticker is auto renew, False otherwise."""
-        if self.owner_info is None:
-            return None
-        return [sticker.is_auto_renew for sticker in self.__iterator__()]
+        return [sticker.is_auto_renew for sticker in self]
 
     @property
-    @_check_sticker
-    def expired_time(self) -> str:
+    def expired_time(self) -> list[Optional[str]]:
         """Returns the sticker's expired time."""
-        if self.owner_info is None:
-            return None
-        return [sticker.expired_time for sticker in self.__iterator__()]
+        return [sticker.expired_time for sticker in self]
 
     @property
-    @_check_sticker
-    def extensions(self) -> StickerExtensions:
+    def extensions(self) -> list[StickerExtensions]:
         """Returns the sticker's extensions."""
-        return StickerExtensions(self.data.get("extensions"))
+        return [sticker.extensions for sticker in self]
 
     @property
-    @_check_sticker
-    def created_time(self) -> str:
+    def created_time(self) -> list[str]:
         """Returns the sticker's created time."""
-        return [sticker.created_time for sticker in self.__iterator__()]
+        return [sticker.created_time for sticker in self]
 
     @property
-    @_check_sticker
-    def is_globally_available(self) -> bool:
+    def is_globally_available(self) -> list[bool]:
         """Returns True if the sticker is globally available, False otherwise."""
-        return [sticker.is_globally_available for sticker in self.__iterator__()]
-    
-    @property   
-    @_check_sticker
-    def restriction_info(self) -> dict:
+        return [sticker.is_globally_available for sticker in self]
+
+    @property
+    def restriction_info(self) -> list[dict[str, Any]]:
         """Returns the sticker's restriction info."""
-        return [sticker.restriction_info for sticker in self.__iterator__()]
+        return [sticker.restriction_info for sticker in self]
 
     @property
-    @_check_sticker
-    def discount_status(self) -> int:
+    def discount_status(self) -> list[int]:
         """Returns the sticker's discount status."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.discount_status for sticker in self.__iterator__()]
+        return [sticker.discount_status for sticker in self]
 
     @property
-    @_check_sticker
-    def owner_uid(self) -> str:
+    def owner_uid(self) -> list[Optional[str]]:
         """Returns the sticker's owner uid."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.owner_uid for sticker in self.__iterator__()]
+        return [sticker.owner_uid for sticker in self]
 
     @property
-    @_check_sticker
-    def owner_type(self) -> int:
+    def owner_type(self) -> list[int]:
         """Returns the sticker's owner type."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.owner_type for sticker in self.__iterator__()]
+        return [sticker.owner_type for sticker in self]
 
     @property
-    @_check_sticker
-    def restrict_type(self) -> int:
+    def restrict_type(self) -> list[int]:
         """Returns the sticker's restrict type."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.restrict_type for sticker in self.__iterator__()]
+        return [sticker.restrict_type for sticker in self]
 
     @property
-    @_check_sticker
-    def restrict_value(self) -> int:
+    def restrict_value(self) -> list[int]:
         """Returns the sticker's restrict value."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.restrict_value for sticker in self.__iterator__()]
+        return [sticker.restrict_value for sticker in self]
 
     @property
-    @_check_sticker
-    def available_duration(self) -> str:
+    def available_duration(self) -> list[Optional[str]]:
         """Returns the sticker's available duration."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.available_duration for sticker in self.__iterator__()]
+        return [sticker.available_duration for sticker in self]
 
     @property
-    @_check_sticker
-    def discount_value(self) -> str:
+    def discount_value(self) -> list[Optional[int]]:
         """Returns the sticker's discount value."""
-        if self.restriction_info is None:
-            return None
-        return [sticker.discount_value for sticker in self.__iterator__()]
+        return [sticker.discount_value for sticker in self]
 
-    def json(self) -> dict:
+    def json(self) -> list[dict[str, Any]]:
         """Returns the sticker's data as a dict."""
         return self.data
