@@ -1,16 +1,14 @@
-from .entities.chat_threads import *
-from .entities.general import *
-from .entities.messages import *
-from .entities.wsevents import *
-from .utilities.generate import *
-from .entities.userprofile import *
-
-from .console import *
-from .socket import *
-from .account import *
-from .context import *
-from .global_client import *
-from .community import *
-from .dispatcher import *
-from .handle_queue import *
-from .utilities.request_handler import *
+from pymino.ext.entities.chat_threads import *
+from pymino.ext.entities.general import *
+from pymino.ext.entities.messages import *
+from pymino.ext.entities.wsevents import *
+from pymino.ext.entities.userprofile import *
+from pymino.ext.utilities.generate import *
+from pymino.ext.utilities.request_handler import *
+from pymino.ext.console import *
+from pymino.ext.socket import *
+from pymino.ext.account import *
+from pymino.ext.context import *
+from pymino.ext.global_client import *
+from pymino.ext.community import *
+from pymino.ext.dispatcher import *
